fix: web server

+++ lazyllm/tools/rag/web.py
@@ -101,7 +101,9 @@ class WebUi:
                 with gr.TabItem('上传文件'):
 
                     def _upload_files(group_name, files):
-
+                        if not files:
+                            gr.Info("请先选择要上传的文件")
+                            return
                         files_to_upload = [
                             ('files', (os.path.basename(file), open(file, 'rb')))
                             for file in files
@@ -135,16 +137,14 @@ class WebUi:
                     def _list_group_files(group_name):
                         file_list = self.list_files_in_group(group_name)
                         values = [[i] + file_list[i][:2] for i in range(len(file_list))]
-                        return gr.update(
-                            value=values
-                        )
+                        return gr.update(value=values)
 
                     select_group = gr.Dropdown(self.list_groups(), label='选择分组')
+                    refresh_btn = gr.Button('刷新')
                     show_list = self.gr_show_list([], list_name=['file_id', 'file_name'])
-                    select_group.change(
-                        fn=_list_group_files, inputs=select_group, outputs=show_list
-                    )
-                    select_group_list.append(select_group)
+
+                    select_group.change(fn=_list_group_files, inputs=select_group, outputs=show_list)
+                    refresh_btn.click(fn=_list_group_files, inputs=select_group, outputs=show_list)
 
                 with gr.TabItem('删除文件'):
 
@@ -154,13 +154,18 @@ class WebUi:
                         return gr.update(choices=file_list)
 
                     select_group = gr.Dropdown(self.list_groups(), label='选择分组')
+                    refresh_btn = gr.Button('刷新文件列表')
                     select_file = gr.Dropdown([], label='选择文件')
-                    select_group.change(
-                        fn=_list_group_files, inputs=select_group, outputs=select_file
-                    )
+
+                    select_group.change(fn=_list_group_files, inputs=select_group, outputs=select_file)
+                    refresh_btn.click(fn=_list_group_files, inputs=select_group, outputs=select_file)
+
                     delete_btn = gr.Button('删除')
 
                     def _delete_file(group_name, select_file):
+                        if not select_file:
+                            gr.Info("请先选择要删除的文件")
+                            return gr.update()
                         file_ids = [select_file.split(',')[0]]
                         gr.Info(self.delete_file(group_name, file_ids))
                         return _list_group_files(group_name)
